// src/graphql/enums.ts
// GraphQL Enums from jewellery-service

export enum JewelleryTypeEnum {
  RING = 'RING',
  ENGAGEMENT_RING = 'ENGAGEMENT_RING',
  WEDDING_BAND = 'WEDDING_BAND',
  BRACELET = 'BRACELET',
  EARRINGS = 'EARRINGS',
  NECKLACE = 'NECKLACE',
  PENDANT = 'PENDANT',
  OTHER = 'OTHER',
}

export enum JewelleryMetalTypeEnum {
  WHITE_GOLD = 'WHITE_GOLD',
  YELLOW_GOLD = 'YELLOW_GOLD',
  ROSE_GOLD = 'ROSE_GOLD',
  PLATINUM = 'PLATINUM',
  STERLING_SILVER = 'STERLING_SILVER',
  PALLADIUM = 'PALLADIUM',
  TUNGSTEN = 'TUNGSTEN',
  TITANIUM = 'TITANIUM',
  CERAMIC = 'CERAMIC',
  STAINLESS_STEEL = 'STAINLESS_STEEL',
  ZIRCONIUM = 'ZIRCONIUM',
  COBALT = 'COBALT',
  DAMASCUS_STEEL = 'DAMASCUS_STEEL',
  TANTALUM = 'TANTALUM',
  SILICONE = 'SILICONE',
  OTHER = 'OTHER',
}

export enum JewelleryMetalKaratEnum {
  KT_9KT = 'KT_9KT',
  KT_10KT = 'KT_10KT',
  KT_12KT = 'KT_12KT',
  KT_14KT = 'KT_14KT',
  KT_18KT = 'KT_18KT',
  KT_22KT = 'KT_22KT',
  KT_925S = 'KT_925S',
  KT_950P = 'KT_950P',
}

export enum JewellerySettingEnum {
  FINISHED = 'FINISHED',
  MOUNT = 'MOUNT',
  SEMI_FINISHED = 'SEMI_FINISHED',
}

export enum JewelleryRingType {
  ENGAGEMENT_RINGS = 'ENGAGEMENT_RINGS',
  WEDDING_BANDS = 'WEDDING_BANDS',
  ETERNITY_BANDS = 'ETERNITY_BANDS',
  COCKTAIL_RINGS = 'COCKTAIL_RINGS',
  STACKABLE_RINGS = 'STACKABLE_RINGS',
  SIGNET_RINGS = 'SIGNET_RINGS',
  PROMISE_RINGS = 'PROMISE_RINGS',
}

export enum JewelleryEngagementRingCategoryEnum {
  ACCENTED = 'ACCENTED',
  BEZEL = 'BEZEL',
  CLASSIC = 'CLASSIC',
  HALO = 'HALO',
  MODERN = 'MODERN',
  SIDE_STONES = 'SIDE_STONES',
  SOLITAIRE = 'SOLITAIRE',
  THREE_STONE = 'THREE_STONE',
  TWO_STONE = 'TWO_STONE',
  VINTAGE_AND_SCULPTURAL = 'VINTAGE_AND_SCULPTURAL',
}

export enum JewelleryEngagementRingFinishEnum {
  BESD_BLAST_MATTE = 'BESD_BLAST_MATTE',
  BRUSHED = 'BRUSHED',
  FILIGREE = 'FILIGREE',
  FLORENTINE = 'FLORENTINE',
  GLASS_BLAST = 'GLASS_BLAST',
  HAMMER = 'HAMMER',
  HIGH_POLISH = 'HIGH_POLISH',
  ICE = 'ICE',
  MATTE = 'MATTE',
  NO_FINISH = 'NO_FINISH',
  ROCK = 'ROCK',
  SATIN = 'SATIN',
  SATIN_HAMMER = 'SATIN_HAMMER',
  SATIN_ROCK = 'SATIN_ROCK',
  SILK = 'SILK',
  STONE = 'STONE',
}

export enum JewelleryClaspTypeEnum {
  BALL = 'BALL',
  BARREL = 'BARREL',
  BOLO = 'BOLO',
  BOX = 'BOX',
  FISHHOOK = 'FISHHOOK',
  HOOK = 'HOOK',
  LOBSTER = 'LOBSTER',
  MAGNETIC = 'MAGNETIC',
  SPRING_RING = 'SPRING_RING',
  TOGGLE = 'TOGGLE',
  SIDE_SAFETY = 'SIDE_SAFETY',
}

export enum CatalogSourceEnum {
  NIVODA = 'NIVODA',
  STULLER = 'STULLER',
}

export enum JewelleryStonePosition {
  CENTER = 'CENTER',
  SIDE = 'SIDE',
  SIGNATURE = 'SIGNATURE',
}

export enum RingStyleTypeEnum {
  SHOWROOM = 'SHOWROOM',
  SHOPIFY = 'SHOPIFY',
  GENERAL = 'GENERAL',
}

export enum Shape {
  ASSCHER = 'ASSCHER',
  BRIOLETTE = 'BRIOLETTE',
  BAGUETTE = 'BAGUETTE',
  BULLET = 'BULLET',
  CALF = 'CALF',
  CUSHION = 'CUSHION',
  CUSHION_ALL = 'CUSHION_ALL',
  CUSHION_BRILLIANT = 'CUSHION_BRILLIANT',
  CUSHION_MODIFIED = 'CUSHION_MODIFIED',
  EMERALD = 'EMERALD',
  EUROPEAN_CUT = 'EUROPEAN_CUT',
  FANCY = 'FANCY',
  FLANDERS = 'FLANDERS',
  HALF_MOON = 'HALF_MOON',
  HEART = 'HEART',
  HEXAGONAL = 'HEXAGONAL',
  KITE = 'KITE',
  LOZENGE = 'LOZENGE',
  MARQUISE = 'MARQUISE',
  OCTAGONAL = 'OCTAGONAL',
  OLD_MINER = 'OLD_MINER',
  OTHER = 'OTHER',
  OVAL = 'OVAL',
  PEAR = 'PEAR',
  PENTAGONAL = 'PENTAGONAL',
  PRINCESS = 'PRINCESS',
  RADIANT = 'RADIANT',
  ROUND = 'ROUND',
  ROSE = 'ROSE',
  SHIELD = 'SHIELD',
  SQ_EMERALD = 'SQ_EMERALD',
  SQ_RADIANT = 'SQ_RADIANT',
  SQUARE = 'SQUARE',
  STAR = 'STAR',
  STEP_CUT_EPAULETTE = 'STEP_CUT_EPAULETTE',
  TAPERED_BAGUETTE = 'TAPERED_BAGUETTE',
  TAPERED_BULLET = 'TAPERED_BULLET',
  TRAPEZOID = 'TRAPEZOID',
  TRIANGULAR = 'TRIANGULAR',
  TRILLIANT = 'TRILLIANT',
  EGG = 'EGG',
}

export enum SortFieldsEnum {
  PRICE = 'PRICE',
}

export enum OrderDirection {
  ASC = 'ASC',
  DESC = 'DESC',
}